Did not stage the rest of the changes.  Restating.
Changed the command "version" to "info".  Currently it only prints out the language version number (which is arbitrary at this point).
Added the commands "add-broker", "console", "get", and "use" to the usage statement, but have yet to implement them.
Completed the implementation of the "create" command.
Anytime a user uses the "disco" command, a check occurs for the ".disco.root" directory in the user's home. If it doesn't exist it creates it.